+++ Example/nvecta-ios-sdk-example-app/nvecta-ios-sdk-example-app/ViewController.swift
@@ -7,18 +7,17 @@
 
 import UIKit
 import NVECTASDK
-import notifyvisitors
+//import notifyvisitors
 
 class ViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
         print("viewDidLoad called")
-        
         NVECTA.shared.delegate = self
         NVECTA.shared.show(userToken: nil, customRule: nil)
        
-        NVECTA.shared.trackEvent(forEventName: "", attributes: [:], ltv: "", scope: 1)
+        NVECTA.shared.trackEvent(forEventName: "home_page_visit", attributes: [:], ltv: "", scope: 1)
 //        notifyvisitors.show(nil, customRule: nil)
         // Do any additional setup after loading the view.
     }
@@ -31,10 +30,7 @@ class ViewController: UIViewController {
 
 extension ViewController: NVECTADelegate {
     func nvectaDidTrackEventResponse(_ response: [String : Any]?) {
-        
-        notifyvisitors.getNotificationCenterData { (notificationsData: [AnyHashable : Any]?) in
-            print("getNotificationCenterData response data = \(notificationsData ?? [:])")
-        }
+        print("[NVECTASDK]-{INFO}: did track event response = \(response ?? [:])")
     }
     
 }